+++ android/history.md
@@ -1,5 +1,9 @@
 # Keyman for Android
 
+## 2019-06-28 11.0.2103 stable
+* Changes:
+  * Allow user to change keyboard from lock screen (#1709)
+
 ## 2019-02-27 11.0.2102 stable
 * Bug fix:
   * Fix crash from language picker trying to show error dialog (#1634)